Publishes the Coveralls report only on AppVeyor
At the moment, there are two tools that can be used to upload
a code coverage report to Coveralls from a Cake script:

- coveralls.net (https://github.com/csmacnz/coveralls.net)
- coveralls.io (https://github.com/coveralls-net/coveralls.net)

While coveralls.net can run as a .NET Core Tool, support for
installing these tools from Cake is still a bit experimental
(see https://www.gep13.co.uk/blog/introducing-cake.dotnettool.module).

So, for the time being, we stick with coveralls.io, which is a Windows
executable requiring the full .NET Framework.

This means that the only CI server where we can reliably run it is
AppVeyor.